// src/app/shared/models/tag.interface.ts
export interface TagInterface {
  created: string;
  name: string;
  nickName: string;
  company: string;
  advertiser: string;
  advertiserProvider: string;
  publisher: string;
  browserStatus: string;
  browser: [];
  deviceTypeStatus: string;
  deviceType: [];
  versionStatus: string;
  version: [];
  countryStatus: string;
  country: [];
  adServerUrl: string;
  statType: string;
  subids: [];
  rotationType: string;
  tagUrls: [];
  initialURL: string;
  _id: string;
  _key: string;
  _rev: string;
}